使用cookie作為登入驗證並做管理者分流

+++ yacht/admin/updateYachtPhoto.aspx.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
+using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Web;
@@ -54,24 +55,69 @@ namespace yacht.admin
             loadPhotos();
             //取得選取項目的值
             string selectImageStr = RadioButtonList1.SelectedValue;
+            if (!string.IsNullOrEmpty(selectImageStr))
+            {
+                //刪除圖片檔案
+                string savePath = Server.MapPath("~/admin/upload/yacht/photo/");
+                File.Delete(savePath + selectImageStr);
 
-            //刪除圖片檔案
-            string savePath = Server.MapPath("~/admin/upload/yacht/photo/");
-            File.Delete(savePath + selectImageStr);
 
-
-            //更新刪除後的圖片名稱 JSON 存入資料庫
-            SqlConnection connection = new SqlConnection(WebConfigurationManager.ConnectionStrings["yachtConnectionString"].ConnectionString);
-            string sql = "DELETE FROM yachtPhoto WHERE(photoName = @photoName)";
-            SqlCommand command = new SqlCommand(sql, connection);
-            command.Parameters.AddWithValue("@photoName", selectImageStr);
-            connection.Open();
-            command.ExecuteNonQuery();
-            connection.Close();
+                //更新刪除後的圖片名稱 JSON 存入資料庫
+                SqlConnection connection = new SqlConnection(WebConfigurationManager.ConnectionStrings["yachtConnectionString"].ConnectionString);
+                string sql = "DELETE FROM yachtPhoto WHERE(photoName = @photoName)";
+                SqlCommand command = new SqlCommand(sql, connection);
+                command.Parameters.AddWithValue("@photoName", selectImageStr);
+                connection.Open();
+                command.ExecuteNonQuery();
+                connection.Close();
+            }
 
             //渲染畫面
             RadioButtonList1.Items.Clear();
             loadPhotos();
+        }
+
+        protected void UploadBtn_Click(object sender, EventArgs e)
+        {
+            //將附件檔案存進目標資料夾
+            string SavePath = Server.MapPath("~/admin/upload/yacht/photo/");
+            string id = Request.QueryString["id"];
+            SqlConnection connection = new SqlConnection(WebConfigurationManager.ConnectionStrings["yachtConnectionString"].ConnectionString);
+            if (photoUpload.HasFile)
+            {
+                DirectoryInfo photo = new DirectoryInfo(SavePath);
+                foreach (var uploadFile in photoUpload.PostedFiles)
+                {
+                    string photoName = Path.GetFileName(uploadFile.FileName);
+                    string[] photoNameArray = photoName.Split('.');
+                    int count = 0;
+                    foreach (var file in photo.GetFiles())
+                    {
+                        if (file.Name.Contains(photoNameArray[0]))
+                        {
+                            count++;
+                        }
+                    }
+                    connection.Open();
+                    photoName = photoNameArray[0] + $"({count + 1})." + photoNameArray[1];
+                    uploadFile.SaveAs(SavePath + photoName);
+                    string attachedSQL = "INSERT INTO yachtPhoto(photoName, yacht_id)VALUES(@photoName, @yacht_id)";
+                    SqlCommand sql2 = new SqlCommand(attachedSQL, connection);
+                    sql2.Parameters.Add("@photoName", SqlDbType.NVarChar);
+                    sql2.Parameters["@photoName"].Value = photoName;
+                    sql2.Parameters.Add("@yacht_id", SqlDbType.Int);
+                    sql2.Parameters["@yacht_id"].Value = id;
+                    sql2.ExecuteNonQuery();
+                    connection.Close();
+                }
+                Response.Redirect(Request.Url.ToString());
+            }
+            else
+            {
+                Label5.Visible = true;
+                Label5.ForeColor = Color.Red;
+                Label5.Text = "Please choose file！";
+            }
         }
     }
 }
